+++ gettsim/interface.py
@@ -1,5 +1,6 @@
 import collections
 import copy
+import pprint
 import textwrap
 from pathlib import Path
 
@@ -38,7 +39,7 @@ def compute_taxes_and_transfers(
     user_columns : str list of str
         Names of columns which are preferred over function defined in the tax and
         transfer system.
-    params : dict
+    params : dict, default None
         A pandas Series or dictionary with user provided parameters. Currently just
         mapping a parameter name to a parameter value, in the future we will need more
         metadata. If parameters have the same name as an existing parameter from the
@@ -80,6 +81,8 @@ def compute_taxes_and_transfers(
     elif isinstance(user_columns, str):
         user_columns = [user_columns]
 
+    params = {} if params is None else params
+
     user_functions = [] if user_functions is None else user_functions
     user_functions = load_functions(user_functions)
 
@@ -109,6 +112,8 @@ def compute_taxes_and_transfers(
             set(data) & {"p_id", "hh_id", "tu_id"}
         )
         data = _dict_subset(data, relevant_columns)
+
+    _fail_if_root_nodes_are_missing(dag, data)
 
     results = execute_dag(dag, data, targets, debug)
 
@@ -374,9 +379,8 @@ def _fail_if_user_columns_are_not_in_functions(
         intro = _format_text_for_cmdline(
             f"""
             You passed the following user column{'' if n_cols == 1 else 's'} which {'is'
-            if n_cols == 1 else 'are'} unnecessary because there {'is' if n_cols == 1
-            else 'are'} no corresponding function{'' if n_cols == 1 else 's'} in the
-            internal or user functions.
+            if n_cols == 1 else 'are'} unnecessary because no functions require them as
+            inputs.
             """
         )
         list_ = create_linewise_printed_list(unnecessary_user_columns)
@@ -481,3 +485,21 @@ def _reorder_columns(results):
     remaining_columns = [i for i in results.columns if i not in sorted_ids]
 
     return results[sorted_ids + remaining_columns]
+
+
+def _fail_if_root_nodes_are_missing(dag, data):
+    missing_nodes = []
+    for node in _root_nodes(dag):
+        if node not in data and "function" not in dag.nodes[node]:
+            missing_nodes.append(node)
+
+    if missing_nodes:
+        formatted = pprint.pformat(missing_nodes)
+        raise ValueError(f"The following data columns are missing.\n\n{formatted}")
+
+
+def _root_nodes(dag):
+    for node in dag.nodes:
+        has_no_parents = len(list(dag.predecessors(node))) == 0
+        if has_no_parents:
+            yield node